Implement the confuse library handling in notifiers
I've tested and verified the following to function:
- Telegram
- Pushover

+++ src/chia_log/handlers/daily_stats/stat_accumulators/search_time_stats.py
@@ -22,7 +22,7 @@ class SearchTimeStats(HarvesterActivityConsumer, StatAccumulator):
 
     def consume(self, obj: HarvesterActivityMessage):
         self._num_measurements += 1
-        self._avg_time_seconds += (obj.search_time_seconds - self._avg_time_seconds) / self._num_measurements
+        self._avg_time_seconds += int((obj.search_time_seconds - self._avg_time_seconds) / self._num_measurements)
         if obj.search_time_seconds > 5:
             self._over_5_seconds += 1
         if obj.search_time_seconds > 15:
